refactor: move findOpenApiDocumentUrl helper to @scalar/import
