--- packages/api-client-app/src/main/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './find-openapi-document-url'
