+++ packages/import/src/resolve.test.ts
@@ -1,24 +1,24 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { findOpenApiDocumentUrl } from './find-openapi-document-url'
+import { resolve } from './resolve'
 
 global.fetch = vi.fn()
 
 function createFetchResponse(data: string) {
   return {
     ok: true,
-    text: () => new Promise((resolve) => resolve(data)),
+    text: () => new Promise((r) => r(data)),
   }
 }
 
-describe('findOpenApiDocumentUrl', () => {
+describe('resolve', () => {
   beforeEach(() => {
     // @ts-expect-error
     global.fetch.mockReset()
   })
 
   it('returns JSON urls', async () => {
-    const result = await findOpenApiDocumentUrl(
+    const result = await resolve(
       'https://cdn.jsdelivr.net/npm/@scalar/galaxy/dist/latest.json',
     )
 
@@ -28,7 +28,7 @@ describe('findOpenApiDocumentUrl', () => {
   })
 
   it('returns YAML urls', async () => {
-    const result = await findOpenApiDocumentUrl(
+    const result = await resolve(
       'https://cdn.jsdelivr.net/npm/@scalar/galaxy/dist/latest.yaml',
     )
 
@@ -36,7 +36,7 @@ describe('findOpenApiDocumentUrl', () => {
       'https://cdn.jsdelivr.net/npm/@scalar/galaxy/dist/latest.yaml',
     )
 
-    const otherResult = await findOpenApiDocumentUrl(
+    const otherResult = await resolve(
       'https://cdn.jsdelivr.net/npm/@scalar/galaxy/dist/latest.yml',
     )
 
@@ -46,15 +46,11 @@ describe('findOpenApiDocumentUrl', () => {
   })
 
   it('finds URLs from sandbox URL', async () => {
-    const result = await findOpenApiDocumentUrl(
-      'https://sandbox.scalar.com/p/dlw8v',
-    )
+    const result = await resolve('https://sandbox.scalar.com/p/dlw8v')
 
     expect(result).toBe('https://sandbox.scalar.com/files/dlw8v/openapi.yaml')
 
-    const otherResult = await findOpenApiDocumentUrl(
-      'https://sandbox.scalar.com/e/dlw8v',
-    )
+    const otherResult = await resolve('https://sandbox.scalar.com/e/dlw8v')
 
     expect(otherResult).toBe(
       'https://sandbox.scalar.com/files/dlw8v/openapi.yaml',
@@ -82,7 +78,7 @@ describe('findOpenApiDocumentUrl', () => {
     // @ts-expect-error Mocking types are missing
     fetch.mockResolvedValue(createFetchResponse(html))
 
-    const result = await findOpenApiDocumentUrl('https://example.com/reference')
+    const result = await resolve('https://example.com/reference')
 
     expect(result).toBe(
       'https://cdn.jsdelivr.net/npm/@scalar/galaxy/dist/latest.yaml',
@@ -110,7 +106,7 @@ describe('findOpenApiDocumentUrl', () => {
     // @ts-expect-error Mocking types are missing
     fetch.mockResolvedValue(createFetchResponse(html))
 
-    const result = await findOpenApiDocumentUrl('https://example.com/reference')
+    const result = await resolve('https://example.com/reference')
 
     expect(result).toBe(
       'https://cdn.jsdelivr.net/npm/@scalar/galaxy/dist/latest.yaml',
@@ -138,7 +134,7 @@ describe('findOpenApiDocumentUrl', () => {
     // @ts-expect-error Mocking types are missing
     fetch.mockResolvedValue(createFetchResponse(html))
 
-    const result = await findOpenApiDocumentUrl('https://example.com/reference')
+    const result = await resolve('https://example.com/reference')
 
     expect(result).toBe('https://example.com/openapi.yaml')
   })
@@ -169,7 +165,7 @@ describe('findOpenApiDocumentUrl', () => {
     // @ts-expect-error Mocking types are missing
     fetch.mockResolvedValue(createFetchResponse(html))
 
-    const result = await findOpenApiDocumentUrl('https://example.com/reference')
+    const result = await resolve('https://example.com/reference')
 
     expect(result).toBe('https://example.com/openapi.yaml')
   })
@@ -192,7 +188,7 @@ describe('findOpenApiDocumentUrl', () => {
     // @ts-expect-error Mocking types are missing
     fetch.mockResolvedValue(createFetchResponse(html))
 
-    const result = await findOpenApiDocumentUrl('https://example.com/reference')
+    const result = await resolve('https://example.com/reference')
 
     expect(result).toBe(
       'https://cdn.jsdelivr.net/npm/@scalar/galaxy/dist/latest.yaml',
@@ -219,7 +215,7 @@ describe('findOpenApiDocumentUrl', () => {
     // @ts-expect-error Mocking types are missing
     fetch.mockResolvedValue(createFetchResponse(html))
 
-    const result = await findOpenApiDocumentUrl('https://example.com/reference')
+    const result = await resolve('https://example.com/reference')
 
     expect(result).toBe(
       'https://cdn.jsdelivr.net/npm/@scalar/galaxy/dist/latest.yaml',
